refactor(autoware_traffic_light_arbiter): rename classes to drop Core suffix
Line up file names with class names: the ROS-free arbitration logic is now the
canonical TrafficLightArbiter in traffic_light_arbiter.{hpp,cpp}, and the thin
ROS node is TrafficLightArbiterNode in traffic_light_arbiter_node.{hpp,cpp}.
Previously the base-named file held the Node while the core carried a Core
suffix, so names were crossed.

Pure rename/move, no behavior change. The registered component becomes
autoware::traffic_light::TrafficLightArbiterNode, but the executable name
(traffic_light_arbiter_node), package name, and launch entry are unchanged, so
downstream launch integration is unaffected.
